Bugfix: Unregister-SecretVault failed if cache had not been created. Get-SecretInfo failed if secret names were not unique.  Changed to using unique ID as the name. This also allows more robust piping of results to Get-Secret.

+++ SecretManagement.BitWarden.Extension/public/Unregister-SecretVault.ps1
@@ -8,5 +8,7 @@ function Unregister-SecretVault
 
     # Removes the bitwarden-ext directory and all contents from the computer.
     $BitwardenExtDir = $IsWindows ? "$env:LocalAppData\Microsoft\PowerShell\secretmanagement\bitwarden-ext" : "$HOME/.secretmanagement/bitwarden-ext"
-    Remove-Item -Path $BitwardenExtDir -Recurse
+    if(Test-Path -Path $BitwardenExtDir -PathType Container) {
+        Remove-Item -Path $BitwardenExtDir -Recurse
+    }
 }
